+++ R/wordstem.R
@@ -72,15 +72,14 @@ wordstem.tokenizedTexts <- function(x, language = "porter") {
 #' \dontshow{
 #' txt <- c(one = "Eating eater eaters eats ate.",
 #'          two = "Taxing taxes taxed my tax return.")
-#' th <- hashTokens(tokenize(toLower(txt)))
+#' th <- tokens(toLower(txt))
 #' wordstem(th)
-#' attr(wordstem(th), "vocabulary")
 #' }
-wordstem.tokenizedTextsHashed <- function(x, language = "porter") {
+wordstem.tokens <- function(x, language = "porter") {
     if (all.equal(attributes(x)$ngrams, 1))
-        vocabulary(x) <- wordstem(vocabulary(x), language = language)
+        types(x) <- wordstem(types(x), language = language)
     else 
-        vocabulary(x) <- wordstem_Ngrams(vocabulary(x), language = language)
+        types(x) <- wordstem_Ngrams(types(x), language = language)
 
     tokens_hashed_recompile(x)
 }